backend/requirements.txt
Flask
Flask-CORS
google-cloud-storage
sqlalchemy
psycopg2-binary
python-dotenv